requirements.txt
django
django-cors-headers
djangorestframework
psychopg2-binary
